# api/openapi.yaml
openapi: "3.0.3"
info:
  title: DoIt API
  version: "1.0.0"
  description: Event-sourced task management API
servers:
  - url: /api/v1
security:
  - cookieAuth: []
paths:
...
  /lists:
    get:
      operationId: listLists
      summary: List all lists
      responses:
        "200":
          description: List of lists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/List"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      operationId: createList
      summary: Create a list
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateListRequest"
      responses:
        "201":
          description: List created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
  /lists/{id}:
    delete:
      operationId: deleteList
      summary: Delete a list
      parameters:
        - $ref: "#/components/parameters/ListId"
      responses:
        "204":
          description: List deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /labels:
    get:
      operationId: listLabels
      summary: List all labels
      responses:
        "200":
          description: Label list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Label"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      operationId: createLabel
      summary: Create a label
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLabelRequest"
      responses:
        "201":
          description: Label created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
  /labels/{id}:
    delete:
      operationId: deleteLabel
      summary: Delete a label
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Label deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: doit_token
  parameters:
...
    ListId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: State conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    IdResponse:
      type: object
      required: [id]
      properties:
        id:
          type: string
          format: uuid
...
    Label:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        colour:
          type: string
        created_at:
          type: string
          format: date-time
    List:
      type: object
      required: [id, name, position, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        colour:
          type: string
        icon:
          type: string
        position:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    CreateListRequest:
      type: object
      required: [name, colour, position]
      properties:
        name:
          type: string
        colour:
          type: string
        icon:
          type: string
        position:
          type: string
    CreateLabelRequest:
      type: object
      required: [name, colour]
      properties:
        name:
          type: string
        colour:
          type: string